requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
pydantic[email]
python-multipart